+++ extensions/common/iam/verifiable-credentials/src/test/java/org/eclipse/edc/iam/verifiablecredentials/StatusListRevocationServiceTest.java
@@ -34,17 +34,17 @@ import org.mockserver.verify.VerificationTimes;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import static org.eclipse.edc.iam.verifiablecredentials.spi.model.statuslist.StatusList2021Credential.STATUS_LIST_CREDENTIAL;
-import static org.eclipse.edc.iam.verifiablecredentials.spi.model.statuslist.StatusList2021Credential.STATUS_LIST_INDEX;
-import static org.eclipse.edc.iam.verifiablecredentials.spi.model.statuslist.StatusList2021Credential.STATUS_LIST_PURPOSE;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.model.statuslist2021.StatusList2021Credential.STATUS_LIST_CREDENTIAL;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.model.statuslist2021.StatusList2021Credential.STATUS_LIST_INDEX;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.model.statuslist2021.StatusList2021Credential.STATUS_LIST_PURPOSE;
 import static org.eclipse.edc.junit.assertions.AbstractResultAssert.assertThat;
 import static org.eclipse.edc.util.io.Ports.getFreePort;
 import static org.mockserver.model.HttpRequest.request;
 
-class StatusList2021RevocationServiceTest {
+class StatusListRevocationServiceTest {
     private static final int NOT_REVOKED_INDEX = 1;
     private static final int REVOKED_INDEX = 2;
-    private final StatusList2021RevocationService revocationService = new StatusList2021RevocationService(new ObjectMapper().registerModule(new JavaTimeModule()),
+    private final StatusListRevocationService revocationService = new StatusListRevocationService(new ObjectMapper().registerModule(new JavaTimeModule()),
             5 * 60 * 1000);
     private ClientAndServer clientAndServer;
