+++ spi/common/verifiable-credentials-spi/src/main/java/org/eclipse/edc/iam/verifiablecredentials/spi/model/bitstringstatuslist/BitstringStatusListStatus.java
@@ -1,0 +1,23 @@
+/*
+ *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+package org.eclipse.edc.iam.verifiablecredentials.spi.model.bitstringstatuslist;
+
+import org.eclipse.edc.iam.verifiablecredentials.spi.model.CredentialStatus;
+
+public class BitstringStatusListStatus {
+    public static BitstringStatusListStatus parse(CredentialStatus status) {
+        return null;
+    }
+}
